requirements.txt
lightning==2.2.4
#torch==2.3.0
hydra-core==1.3.2
gin-config==0.5.0